[APPC-1634] Moved listener creation to fragment to respect MVP architecture
